fix: only importing mlx related modules if on macos
Also moved mlx related code to mlx_api.py from api.py.
This is done for now so that the sever and cli runs on linux too.
We will refactor this with better architecture as part of 0.x.x refactoring
