+++ server/pyrightconfig.json
@@ -1,0 +1,5 @@
+{
+  "venvPath": ".",
+  "venv": ".venv"
+}
+
